Pirandello - Six Characters in Search of an Author

This dataset contains the dialogues of the 1921 play Six Characters in Search of an Author by the Italian writer (and Nobel laureate) Luigi Pirandello. The Spanish version contained here was translated in 1926 by Félix Azzati and is currently available at Wikipedia.

Usage

Pirandello

Format

a data.frame object with the following variables:

Order numeric; The order of the speech in the play.

Personaje character; The character responsible for the speech.

Action character; Description of actions, emotions or gestures of the character immediately before his or her speech.

text character; The text of the speech.

Source

Wikipedia: https://es.wikisource.org/wiki/Seis_personajes_en_busca_de_autor.

Examples

# some operations on the corpus
summary(Pirandello)
#>      Order        Personaje            Accion              text          
#>  Min.   :  1.0   Length:683         Length:683         Length:683        
#>  1st Qu.:171.5   Class :character   Class :character   Class :character  
#>  Median :342.0   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :342.0                                                           
#>  3rd Qu.:512.5                                                           
#>  Max.   :683.0                                                           
head(Pirandello)
#>    Order             Personaje Accion
#> 8      1 Descripcion de escena   <NA>
#> 12     2 Descripcion de escena   <NA>
#> 16     3 Descripcion de escena   <NA>
#> 19     4 Descripcion de escena   <NA>
#> 22     5 Descripcion de escena   <NA>
#> 25     6 Descripcion de escena   <NA>
#>                                                                                                                                                                                                                                                                                                                                                             text
#> 8                                                                                                                                                                                                                                                                                                              Es de día, en el escenario de un teatro de verso.
#> 12 NOTA BENE. La Comedia no tiene actos ni escenas. La representación se interrumpirá una primera vez, sin bajar el telón, cuando el Director de la Compañía y el primero de los Personajes se retiran para concertar la escenificación, desalojando entonces los actores el escenario; y una segunda vez, cuando por error, el maquinista dejará caer el telón.
#> 16                                                                                                                  Al entrar los espectadores en el teatro, hallarán levantado el telón, casi obscuro y vacio el escenario, como durante el dia, sin bastidores ni decorado, para que desde el principio se reciba la impresión de un espectáculo no preparado.
#> 19                                                                                                                                                                                                                                                                                                        La concha del apuntador, estará a un lado del boquete.
#> 22                                                                                                                                                                                                                                 Al otro lado, cerca del proscenio, una mesita y una butaca con el respaldo hacia el público, para el Director de la Compañía.
#> 25                                                                                                                                                                                                                                   Otras dos mesitas, una más grande y otra más pequeña, con algunas sillas en torno, por si son necesarias durante el ensayo.