Brazilian Presidents' Inaugural Speeches

This dataset contains 37 Inaugural Speeches from 1889 to 2023.

Usage

bra.inaugural

Format

a quanteda corpus object with the following variables:

doc_id character; Document ID

text character; Text of the inaugural speech

President character; Name of the President who gave the speech

Legislature numeric; The number of the legislature

Date date; Date of the Speech

Military logical; Military President

Party character; Political Party of the President

Interrupted logical; Was the administration interrupted?

Interrupt.Cause character; Cause of the interruption

link_photo character; URL for the photos of the President

Source

Presidencia.gov.br and Bonfim (2005).

Examples

# some operations on the corpus
summary(bra.inaugural)
#> Corpus consisting of 37 documents, showing 37 documents:
#> 
#>                                  Text Types Tokens Sentences Year
#>         01. Deodoro da Fonseca (1889)   193    367         9 1889
#>           02. Floriano Peixoto (1891)   389    835        22 1891
#>         03. Prudente de Moraes (1894)   637   1438        26 1894
#>               04. Campos Sales (1898)  1417   4053       109 1898
#>            05. Rodrigues Alves (1902)   857   2112        55 1902
#>              06. Affonso Penna (1906)  1470   3952       104 1906
#>          07. Hermes da Fonseca (1910)  1138   2966        51 1910
#>             08. Wenceslau Brás (1914)  1526   4285       131 1914
#>            09. Epitácio Pessoa (1919)  1420   4094       113 1919
#>           10. Arthur Bernardes (1922)   319    586        13 1922
#>            11. Washington Luís (1926)   256    505        19 1926
#>             12. Getúlio Vargas (1930)   789   1753        44 1930
#>             13. Getúlio Vargas (1934)   592   1175        47 1934
#>             14. Getúlio Vargas (1937)  1488   3924       102 1937
#>        15. Eurico Gaspar Dutra (1946)   421   1310        25 1946
#>             16. Getúlio Vargas (1951)   604   1310        51 1951
#>       17. Juscelino Kubitschek (1956)   322    711        23 1956
#>              18. Jânio Quadros (1961)   386    737        25 1961
#>               19. João Goulart (1961)   461   1191        49 1961
#>             20. Castelo Branco (1964)   512   1151        37 1964
#>              21. Costa e Silva (1967)   639   1472        54 1967
#>   22. Emílio Garrastazu Médici (1969)   766   2275        85 1969
#>             23. Ernesto Geisel (1974)   379    715        17 1974
#>    24. João Batista Figueiredo (1979)   756   1754        82 1979
#>                25. José Sarney (1985)  1559   4475       261 1985
#>             26. Tancredo Neves (1985)  1198   3282       112 1985
#>    27. Fernando Collor de Melo (1990)  2076   6664       261 1990
#>              28. Itamar Franco (1992)   339    700        34 1992
#>  29. Fernando Henrique Cardoso (1995)  1297   3729       163 1995
#>  30. Fernando Henrique Cardoso (1999)  1112   2976       173 1999
#>  31. Luis Inácio Lula da Silva (2003)  1530   4380       147 2003
#>  32. Luis Inácio Lula da Silva (2007)  1455   4282       201 2007
#>             33. Dilma Rousseff (2011)  1343   4305       177 2011
#>             34. Dilma Rousseff (2015)  1582   5122       195 2015
#>               35. Michel Temer (2016)  1089   3175       138 2016
#>             36. Jair Bolsonaro (2019)   580   1298        60 2019
#>  37. Luis Inácio Lula da Silva (2023)  1213   3318       147 2023
#>                  President       Date Military    Party Interrupted
#>         Deodoro da Fonseca 1889-11-16      Yes Military          No
#>           Floriano Peixoto 1891-11-23      Yes Military          No
#>         Prudente de Moraes 1894-11-15       No      PRF          No
#>               Campos Sales 1898-11-15       No      PRP          No
#>            Rodrigues Alves 1902-11-15       No      PRP          No
#>              Affonso Penna 1906-11-15       No      PRM         Yes
#>          Hermes da Fonseca 1910-11-15      Yes      PRC          No
#>             Wenceslau Brás 1914-11-15       No      PRM         Yes
#>            Epitácio Pessoa 1919-09-03       No      PRM          No
#>           Arthur Bernardes 1922-11-16       No      PRM          No
#>            Washington Luís 1926-11-16       No      PRP         Yes
#>             Getúlio Vargas 1930-10-24       No No party          No
#>             Getúlio Vargas 1934-07-21       No No party          No
#>             Getúlio Vargas 1937-11-10       No No party         Yes
#>        Eurico Gaspar Dutra 1946-01-31      Yes      PSD          No
#>             Getúlio Vargas 1951-01-31       No      PTB         Yes
#>       Juscelino Kubitschek 1956-01-31       No      PSD          No
#>              Jânio Quadros 1961-01-31       No      PTN         Yes
#>               João Goulart 1961-09-08       No      PTB         Yes
#>             Castelo Branco 1964-04-15      Yes Military          No
#>              Costa e Silva 1967-03-15      Yes Military          No
#>   Emílio Garrastazu Médici 1969-10-30      Yes Military          No
#>             Ernesto Geisel 1974-03-15      Yes Military          No
#>    João Batista Figueiredo 1979-03-15      Yes Military          No
#>                José Sarney 1985-03-15       No     PMDB          No
#>             Tancredo Neves 1985-07-22       No     PMDB         Yes
#>    Fernando Collor de Melo 1990-03-15       No      PRN         Yes
#>              Itamar Franco 1992-10-02       No     PMDB          No
#>  Fernando Henrique Cardoso 1995-01-01       No     PSDB          No
#>  Fernando Henrique Cardoso 1999-01-01       No     PSDB          No
#>  Luis Inácio Lula da Silva 2003-01-01       No       PT          No
#>  Luis Inácio Lula da Silva 2007-01-01       No       PT          No
#>             Dilma Rousseff 2011-01-01       No       PT          No
#>             Dilma Rousseff 2015-01-01       No       PT         Yes
#>               Michel Temer       <NA>       No     PMDB          No
#>             Jair Bolsonaro 2019-01-01       No      PSL          No
#>  Luis Inácio Lula da Silva 2023-01-01       No       PT        <NA>
#>  Interrupt.Cause Elected                link_photo
#>               No      No https://shorturl.at/bjUY0
#>               No      No https://shorturl.at/luzEG
#>               No     Yes https://shorturl.at/bdgAR
#>               No     Yes https://shorturl.at/buLQ5
#>               No     Yes https://shorturl.at/lwCN3
#>            Death     Yes https://shorturl.at/gwGX2
#>               No     Yes https://shorturl.at/bcoK2
#>            Death     Yes https://shorturl.at/gluJN
#>               No      No https://shorturl.at/JU147
#>               No     Yes https://shorturl.at/wAHJZ
#>             Coup     Yes https://shorturl.at/tRZ27
#>               No      No https://shorturl.at/antuC
#>               No      No https://shorturl.at/antuC
#>             Coup      No https://shorturl.at/antuC
#>               No     Yes https://shorturl.at/fkoEH
#>          Suicide     Yes https://shorturl.at/brGWX
#>               No     Yes https://shorturl.at/nHP59
#>        Renounced     Yes https://shorturl.at/bDGLM
#>             Coup      No https://shorturl.at/asIT4
#>               No      No https://shorturl.at/aoqsS
#>               No      No https://shorturl.at/hrX14
#>               No      No https://shorturl.at/lmAHX
#>               No      No https://shorturl.at/oszGT
#>               No      No https://shorturl.at/bikOR
#>               No      No https://shorturl.at/fqzG5
#>            Death      No https://shorturl.at/mAWX5
#>      Impeachment     Yes https://shorturl.at/agzWY
#>               No      No https://shorturl.at/euM45
#>               No     Yes https://shorturl.at/ghNX8
#>               No     Yes https://shorturl.at/ghNX8
#>               No     Yes https://shorturl.at/dsAM3
#>               No     Yes https://shorturl.at/dsAM3
#>               No     Yes https://shorturl.at/lpvB6
#>      Impeachment     Yes https://shorturl.at/lpvB6
#>               No      No https://shorturl.at/pH489
#>               No     Yes https://shorturl.at/bsu56
#>             <NA>     Yes https://shorturl.at/ijkyY
#> 
bra.inaugural
#> Corpus consisting of 37 documents and 9 docvars.
#> 01. Deodoro da Fonseca (1889) :
#> "Concidadãos – O povo, o exército e a armada nacional, em per..."
#> 
#> 02. Floriano Peixoto (1891) :
#> "Tendo assumido o governo do Estado nos termos da Constituiçã..."
#> 
#> 03. Prudente de Moraes (1894) :
#> "Á Nação Brasileira* Assumindo hoje a Presidência da Repúblic..."
#> 
#> 04. Campos Sales (1898) :
#> "Ao assumir o governo da República, cheio de confiança nos po..."
#> 
#> 05. Rodrigues Alves (1902) :
#> "À Nação Assumindo hoje o cargo de Presidente da República, p..."
#> 
#> 06. Affonso Penna (1906) :
#> "À NAÇÃO Ao Assumir a Suprema Magistratura da Republica, no p..."
#> 
#> [ reached max_ndoc ... 31 more documents ]