Political Corruption in Spain

This dataset contains the transcription of 7 focus groups on political corruption in Spain. The study was performed by the Centro de Investigaciones Sociológicas (CIS) under the number E2863 in March, 2011.

Usage

cis.corrupt

Format

a data.frame object with the following variables:

Persona character; The person speaking.

text character; Text containing the intervention of each participant.

Order integer; Order in the overall interventions.

Tipo.Persona character; Tipe of participant: Moderator, Female, or Male.

Estudio character; Study number: E2863.

Grupo.Discusion integer; Number of the focus group.

Ciudad character; Name of the city where the focus group took place.

Fecha date; Date of the focus group.

Grupo.Demografico character; Name of the demographic group: "Housewives", "Businessmen","University Students", "Public Servants", "Occupied - Young", "Workers", and "Liberal Professionals".

Desc.Grupo character; Detailed description of each demographic group.

Edad character; Age interval of participants.

Source

CIS (2011).

Examples

# some operations on the data
summary(cis.corrupt)
#>    Persona              text               Order       Tipo.Persona      
#>  Length:4445        Length:4445        Min.   :  1.0   Length:4445       
#>  Class :character   Class :character   1st Qu.:159.0   Class :character  
#>  Mode  :character   Mode  :character   Median :318.0   Mode  :character  
#>                                        Mean   :335.6                     
#>                                        3rd Qu.:495.0                     
#>                                        Max.   :837.0                     
#>    Estudio          Grupo.Discusion    Ciudad             Fecha          
#>  Length:4445        Min.   :1.000   Length:4445        Length:4445       
#>  Class :character   1st Qu.:2.000   Class :character   Class :character  
#>  Mode  :character   Median :4.000   Mode  :character   Mode  :character  
#>                     Mean   :3.842                                        
#>                     3rd Qu.:6.000                                        
#>                     Max.   :7.000                                        
#>  Grupo.Demografico   Desc.Grupo            Edad          
#>  Length:4445        Length:4445        Length:4445       
#>  Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character  
#>                                                          
#>                                                          
#>                                                          
head(cis.corrupt)
#>   Persona
#> 1     Mod
#> 2      M1
#> 3     Mod
#> 4      M1
#> 5     Mod
#> 6      M1
#>                                                                                                                                                                                                                                                                                                                                                                                              text
#> 1  Bueno, en primer lugar queria daros la bienvenida y agradeceros que os hayais animado a pasar aqui una tarde conmigo de trabajo, por decirlo. Os queria pedir, por favor, si no teneis ningun problema, permiso para grabar la sesion porque asi puedo estar mas tranquila, pendiente de la conversacion y no tener que estar tomando ninguna nota. Bueno, os voy a explicar un poco de que va
#> 2                                                                                                                                                                                                                                                                                                         Perdona, es para preparar una entrevista de trabajo, sera un segundo. Esta sin volumen.
#> 3                                                                                                                                                                                                                                                                                                                                                                               ?Te van a llamar?
#> 4                                                                                                                                                                                                                                                                                                                                                                          Si, no salgo siquiera.
#> 5                                                                                                                                                                                                                                                                                                                                                                        Bueno, lo que prefieras.
#> 6                                                                                                                                                                                                                                                                                                                                                                  "Si a las seis, seis y media".
#>   Order Tipo.Persona Estudio Grupo.Discusion   Ciudad      Fecha
#> 1     1    Moderador   E2863               1 Valencia 2011-03-10
#> 2     2        Mujer   E2863               1 Valencia 2011-03-10
#> 3     3    Moderador   E2863               1 Valencia 2011-03-10
#> 4     4        Mujer   E2863               1 Valencia 2011-03-10
#> 5     5    Moderador   E2863               1 Valencia 2011-03-10
#> 6     6        Mujer   E2863               1 Valencia 2011-03-10
#>   Grupo.Demografico
#> 1      Amas de casa
#> 2      Amas de casa
#> 3      Amas de casa
#> 4      Amas de casa
#> 5      Amas de casa
#> 6      Amas de casa
#>                                                                                                                               Desc.Grupo
#> 1 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#> 2 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#> 3 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#> 4 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#> 5 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#> 6 Grupo de amas de casa, 50 años a 60 (compañeras de empleados y funcionarios con empleados bajo su responsabilidad, de técnicos medios)
#>    Edad
#> 1 50-60
#> 2 50-60
#> 3 50-60
#> 4 50-60
#> 5 50-60
#> 6 50-60